refctor: extract web.py
This is a surprisingly hefty set of changes to extract the web service
component into web.py.

- For now, we pass the argparse arguments object around diretly to components.
We'll refactor that later.
- Remove the mechanism that renders docs for web service using subprocess.
We'll work out how to do hot module reload later.

I have no doubt this patch breaks some things, but it's the first step to
getting to the point where we can unit test the web service component. Let's
look at the structure first, then consolidate.
