+++ pdoc/templates/html.mako
@@ -9,7 +9,7 @@
   import pygments.formatters
   import pygments.lexers
 
-  import pdoc
+  import pdoc.doc
 
   # From language reference, but adds '.' to allow fully qualified names.
   pyident = re.compile('^[a-zA-Z_][a-zA-Z0-9_.]+$')
@@ -96,10 +96,10 @@
            '' if base == '..' else
            base)
     if m.is_package():
-      index = pdoc.html_package_name
+      index = pdoc.doc.html_package_name
       url = url + '/' + index if url else index
     else:
-      url += pdoc.html_module_suffix
+      url += pdoc.doc.html_module_suffix
     return link_prefix + url
 
   def external_url(refname):
@@ -133,12 +133,12 @@
     considered external and no link is used.
     """
     d = module.find_ident(refname)
-    if isinstance(d, pdoc.External):
+    if isinstance(d, pdoc.doc.External):
       if is_external_linkable(refname):
         return d.refname, external_url(d.refname)
       else:
         return None, None
-    if isinstance(d, pdoc.Module):
+    if isinstance(d, pdoc.doc.Module):
       return d.refname, module_url(d)
     if module.is_public(d.refname):
       return d.name, '#%s' % d.refname
@@ -180,7 +180,7 @@
     <div class="desc">${docstring | mark}</div>
   % endif
   % endif
-  % if not isinstance(d, pdoc.Module):
+  % if not isinstance(d, pdoc.doc.Module):
   <div class="source_cont">${show_source(d)}</div>
   % endif
 </%def>
@@ -476,7 +476,7 @@
   <div class="clear"> </div>
   <footer id="footer">
     <p>
-      Generated by <a href="https://github.com/mitmproxy/pdoc">pdoc ${pdoc.__version__}</a>
+      Generated by <a href="https://github.com/mitmproxy/pdoc">pdoc ${pdoc.doc.__version__}</a>
     </p>
   </footer>
 </div>
